update shipOrder(), employeeInterface(), add displayUnshippedOrders(), getShippingSpeedPriority()
